Fix incorrect kernel argument alignment for TruncatedNormalKernel

+++ tensorflow/core/kernels/parameterized_truncated_normal_op_gpu.cu.cc
@@ -211,7 +211,8 @@ struct TruncatedNormalFunctor<GPUDevice, T> {
         gen, output.data(), num_batches, samples_per_batch, num_elements,
         means.data(), means.dimension(0) == 1, stddevs.data(),
         stddevs.dimension(0) == 1, minvals.data(), minvals.dimension(0) == 1,
-        maxvals.data(), maxvals.dimension(0) == 1, kMaxIterations);
+        maxvals.data(), maxvals.dimension(0) == 1,
+        static_cast<int64>(kMaxIterations));
   };
 };
